datafeeder: refactor packages service.batch.* to batch.*
