+++ datafeeder/src/main/java/org/georchestra/datafeeder/batch/analysis/DatasetUploadStateItemReader.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License along with
  * geOrchestra.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.georchestra.datafeeder.service.batch.analysis;
+package org.georchestra.datafeeder.batch.analysis;
 
 import java.util.Iterator;
 import java.util.List;
