+++ datafeeder/src/main/java/org/georchestra/datafeeder/batch/publish/MockOWSPublicationService.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License along with
  * geOrchestra.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.georchestra.datafeeder.service.batch.publish;
+package org.georchestra.datafeeder.batch.publish;
 
 import java.net.URI;
 import java.util.Objects;
