+++ datafeeder/src/main/java/org/georchestra/datafeeder/batch/publish/task/DataImportTasklet.java
@@ -16,11 +16,11 @@
  * You should have received a copy of the GNU General Public License along with
  * geOrchestra.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.georchestra.datafeeder.service.batch.publish.task;
+package org.georchestra.datafeeder.batch.publish.task;
 
 import java.util.UUID;
 
-import org.georchestra.datafeeder.service.batch.publish.PublishingBatchService;
+import org.georchestra.datafeeder.batch.publish.PublishingBatchService;
 import org.springframework.batch.core.StepContribution;
 import org.springframework.batch.core.scope.context.ChunkContext;
 import org.springframework.batch.core.step.tasklet.Tasklet;
