+++ datafeeder/src/main/java/org/georchestra/datafeeder/repository/DatasetUploadStateRepository.java
@@ -21,8 +21,8 @@ package org.georchestra.datafeeder.repository;
 import java.util.List;
 import java.util.UUID;
 
-import org.georchestra.datafeeder.model.JobStatus;
 import org.georchestra.datafeeder.model.DatasetUploadState;
+import org.georchestra.datafeeder.model.JobStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
