+++ datafeeder/src/main/java/org/georchestra/datafeeder/service/DataUploadService.java
@@ -26,12 +26,12 @@ import java.util.Optional;
 import java.util.UUID;
 
 import org.georchestra.datafeeder.api.FileUploadApiController;
-import org.georchestra.datafeeder.model.JobStatus;
+import org.georchestra.datafeeder.batch.analysis.DataUploadAnalysisService;
 import org.georchestra.datafeeder.model.BoundingBoxMetadata;
 import org.georchestra.datafeeder.model.DataUploadJob;
 import org.georchestra.datafeeder.model.DatasetUploadState;
+import org.georchestra.datafeeder.model.JobStatus;
 import org.georchestra.datafeeder.repository.DataUploadJobRepository;
-import org.georchestra.datafeeder.service.batch.analysis.DataUploadAnalysisService;
 import org.opengis.feature.simple.SimpleFeature;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Async;
